book_inventory/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
drf-yasg==1.20.0
Pillow
django-filter==2.4.0
dj-rest-auth[with_social]==2.1.11
psycopg2>=2.8
pytest